+++ crates/sdk-core/tests/integ_tests/update_tests.rs
@@ -672,7 +672,7 @@ async fn update_with_local_acts() {
     impl UpdateWithLocalActsWf {
         #[run]
         async fn run(ctx: &mut WorkflowContext<Self>) -> WorkflowResult<()> {
-            ctx.wait_condition(|s| s.done).await;
+            ctx.wait_condition(|s| s.done).await?;
             Ok(())
         }
 
@@ -731,10 +731,7 @@ async fn update_with_local_acts() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, run);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
 }
 
 #[tokio::test]
@@ -797,10 +794,7 @@ async fn update_rejection_sdk() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, run);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
 }
 
 #[tokio::test]
@@ -854,10 +848,7 @@ async fn update_fail_sdk() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, run);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
 }
 
 #[tokio::test]
@@ -878,7 +869,7 @@ async fn unknown_update_rejected_sdk() {
     impl UnknownUpdateRejectedSdkWf {
         #[run]
         async fn run(ctx: &mut WorkflowContext<Self>) -> WorkflowResult<()> {
-            ctx.wait_condition(|s| s.done).await;
+            ctx.wait_condition(|s| s.done).await?;
             Ok(())
         }
 
@@ -951,7 +942,7 @@ async fn update_timer_sequence() {
     impl UpdateTimerSequenceWf {
         #[run]
         async fn run(ctx: &mut WorkflowContext<Self>) -> WorkflowResult<()> {
-            ctx.wait_condition(|s| s.done).await;
+            ctx.wait_condition(|s| s.done).await?;
             Ok(())
         }
 
@@ -991,10 +982,7 @@ async fn update_timer_sequence() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, run);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
 }
 
 #[tokio::test]
@@ -1064,10 +1052,7 @@ async fn task_failure_during_validation() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, run);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
     // Verify we did not spam task failures. There should only be one.
     let history = starter.get_history().await;
     assert_eq!(
@@ -1138,10 +1123,7 @@ async fn task_failure_after_update() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, run);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
 }
 
 static BARR: LazyLock<Barrier> = LazyLock::new(|| Barrier::new(2));
@@ -1179,7 +1161,7 @@ async fn worker_restarted_in_middle_of_update() {
     impl WorkerRestartedInMiddleOfUpdateWf {
         #[run]
         async fn run(ctx: &mut WorkflowContext<Self>) -> WorkflowResult<()> {
-            ctx.wait_condition(|s| s.done).await;
+            ctx.wait_condition(|s| s.done).await?;
             Ok(())
         }
 
@@ -1267,10 +1249,7 @@ async fn worker_restarted_in_middle_of_update() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, run, stopper);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
 }
 
 #[tokio::test]
@@ -1293,7 +1272,9 @@ async fn update_after_empty_wft() {
         #[run]
         async fn run(ctx: &mut WorkflowContext<Self>) -> WorkflowResult<()> {
             let sig_handle = async {
-                ctx.wait_condition(|s| s.signal_received).await;
+                ctx.wait_condition(|s| s.signal_received)
+                    .await
+                    .expect("workflow was not cancelled");
                 ACT_STARTED.store(true, Ordering::Release);
                 let _ = ctx
                     .execute_activity(
@@ -1364,10 +1345,7 @@ async fn update_after_empty_wft() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, runner);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
 }
 
 #[tokio::test]
@@ -1389,7 +1367,7 @@ async fn update_lost_on_activity_mismatch() {
         async fn run(ctx: &mut WorkflowContext<Self>) -> WorkflowResult<()> {
             ctx.state_mut(|s| s.can_run = 1);
             for _ in 1..=3 {
-                ctx.wait_condition(|s| s.can_run > 0).await;
+                ctx.wait_condition(|s| s.can_run > 0).await?;
                 let _ = ctx
                     .execute_activity(
                         StdActivities::echo,
@@ -1442,8 +1420,5 @@ async fn update_lost_on_activity_mismatch() {
         worker.run_until_done().await.unwrap();
     };
     join!(update, runner);
-    handle
-        .fetch_history_and_replay(worker.inner_mut())
-        .await
-        .unwrap();
+    handle.fetch_history_and_replay(&mut worker).await.unwrap();
 }