+++ crates/sdk-core/tests/integ_tests/worker_versioning_tests.rs
@@ -327,7 +327,7 @@ impl ContinueAsNewAutoUpgradeV1 {
             return Ok("v1.0".to_string());
         }
         ctx.wait_condition(|state| state.should_continue_as_new)
-            .await;
+            .await?;
         assert!(ctx.target_worker_deployment_version_changed());
         let mut options = ContinueAsNewOptions::default();
         options.initial_versioning_behavior = Some(ContinueAsNewVersioningBehavior::AutoUpgrade);
@@ -455,7 +455,7 @@ impl ContinueAsNewUseRampingVersionV1 {
             return Ok("v1.0".to_string());
         }
         ctx.wait_condition(|state| state.should_continue_as_new)
-            .await;
+            .await?;
         let mut options = ContinueAsNewOptions::default();
         options.initial_versioning_behavior =
             Some(ContinueAsNewVersioningBehavior::UseRampingVersion);