+++ crates/sdk/examples/cancellation/workflows.rs
@@ -2,7 +2,8 @@
 use std::time::Duration;
 use temporalio_macros::{activities, workflow, workflow_methods};
 use temporalio_sdk::{
-    ActivityOptions, CancellableFuture, WorkflowContext, WorkflowResult,
+    ActivityExecutionError, ActivityOptions, WorkflowCancellationToken, WorkflowContext,
+    WorkflowResult,
     activities::{ActivityContext, ActivityError},
 };
 
@@ -44,30 +45,32 @@ pub struct CancellationWorkflow;
 impl CancellationWorkflow {
     #[run]
     pub async fn run(ctx: &mut WorkflowContext<Self>, _input: ()) -> WorkflowResult<String> {
-        let mut activity_fut = ctx.execute_activity(
-            CancellationActivities::long_running_activity,
-            (),
-            activity_opts(),
-        );
-
-        temporalio_sdk::workflows::select! {
-            result = &mut activity_fut => {
-                let value = result?;
-                Ok(value)
-            }
-            reason = ctx.cancelled() => {
-                activity_fut.cancel();
-
+        let result = ctx
+            .execute_activity(
+                CancellationActivities::long_running_activity,
+                (),
+                activity_opts(),
+            )
+            .await;
+        match result {
+            Ok(value) => Ok(value),
+            Err(ActivityExecutionError::Cancelled(_)) => {
+                let reason = ctx.cancellation_token().reason().unwrap_or_default();
                 let cleanup_result = ctx
                     .execute_activity(
                         CancellationActivities::cleanup,
                         (),
-                        ActivityOptions::start_to_close_timeout(Duration::from_secs(10)),
+                        ActivityOptions::with_start_to_close_timeout(Duration::from_secs(10))
+                            // Use a cancellation token disconected from the workflow cancellation to
+                            // ensure cleanup activity is run.
+                            .cancellation_token(WorkflowCancellationToken::new())
+                            .build(),
                     )
                     .await?;
 
                 Ok(format!("Cancelled (reason={reason}), {cleanup_result}"))
             }
+            Err(err) => Err(err.into()),
         }
     }
 }